refactor(sandbox): split host_check.rs into per-check modules
Split the monolithic host_check.rs (580 lines) into a checks/ directory
following the existing pattern in runner/ and host/ subdirectories:

- checks/mod.rs — Status, Outcome, Report types, run(), plural()
- checks/linux/mod.rs — declares all 8 Linux-gated check submodules
- checks/linux/kernel_version.rs — check_kernel_version() + parse_kernel_version()
- checks/linux/user_namespaces.rs — check_user_namespaces()
- checks/linux/cgroup2.rs — check_cgroup_v2()
- checks/linux/brokkr_slice.rs — check_brokkr_slice()
- checks/linux/subtree_controllers.rs — check_subtree_controllers() + parse_subtree_controllers()
- checks/linux/seccomp.rs — check_seccomp()
- checks/linux/memory_peak.rs — check_memory_peak()
- checks/linux/setgroups.rs — check_setgroups()

Updated brokkr-worker to use `brokkr_sandbox::checks` instead of
`brokkr_sandbox::host_check`. All tests pass, clippy clean, fmt clean.

Addresses jackthepunished/brokkr#28.
